+++ src/utils/isMacOS.js
@@ -1,1 +1,0 @@
-export default () => window.navigator.platform.indexOf('Mac') > -1;